Maintain per-type high water process count in cxy_ctl.
Adds a cxy_ctl:high_water function, which returns the maximum number of
concurrent processes run by cxy_ctl:execute_task family of functions.
The one-argument form just returns the value, while the two-argument
form also atomically resets the high-water mark.

The high-water value is maintained as a new sixth field in the Task_Type
records in the cxy_ctl ets table.  The negative value is stored in the
table, because that allows us to play a trick with ets:update_counter in
cxy_ctl:incr_active_procs/1 to atomically maintain the maximum value.
There is a related ets:update_counter trick in cxy_ctl:high_water/2 to
atomically read and reset the high-water value.