Bump version to 1.0.0
This is not the "official worthy" 1.0 release, but I really needed to stop artificially piling up merges on 0.9.9. The trigger was the outside contribution of PR #58 for high watermark on cxy_ctl. The official release probably won't happen until 2.0.0 since there is much to update on documentation, diagrams, and an official explanation and consistent approach to the patterns used in this library.