Delete cache before create so cxy_cache_fsm restart succeeds
cxy_cache_fsm exists solely to own the ets generational
tables for a cache. If it stops, the cache tables will be
automatically deleted. So the stop clause needed to delete
the cache metadata from the ets table it doesn't own.

A more complicated situation arises when cxy_cache_fsm
restarts. It may have died in a way that prevented terminate
from running, so the ets tables are gone, but the metadata
is still present. The sledgehammer approach is to use the
cxy_cache_sup:start_link function to delete a cache before
it instantiates a new cxy_cache_fsm instance.

If this behavior is not wanted, call cxy_cache_fsm:start_link
directly rather than via cxy_cache_sup, or roll your own
cxy_cache_sup to change the initialization sequence.